add contentview
